+++ migrations/Version20260208191001.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DoctrineMigrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Auto-generated Migration: Please modify to your needs!
+ */
+final class Version20260208191001 extends AbstractMigration
+{
+    public function getDescription(): string
+    {
+        return '';
+    }
+
+    public function up(Schema $schema): void
+    {
+        // this up() migration is auto-generated, please modify it to your needs
+        $this->addSql('ALTER TABLE comic_series ADD deleted_at DATETIME DEFAULT NULL');
+        $this->addSql('CREATE INDEX idx_comic_series_deleted_at ON comic_series (deleted_at)');
+    }
+
+    public function down(Schema $schema): void
+    {
+        // this down() migration is auto-generated, please modify it to your needs
+        $this->addSql('DROP INDEX idx_comic_series_deleted_at ON comic_series');
+        $this->addSql('ALTER TABLE comic_series DROP deleted_at');
+    }
+}